atcoder/joi2007ho/c_test.py
from c import resolve
import sys
from io import StringIO
import unittest

class TestClass(unittest.TestCase):
    maxDiff = None
    def assertIO(self, input, output):
        stdout, stdin = sys.stdout, sys.stdin
        sys.stdout, sys.stdin = StringIO(), StringIO(input)
        resolve()
        sys.stdout.seek(0)
        out = sys.stdout.read()[:-1]
        sys.stdout, sys.stdin = stdout, stdin
        self.assertEqual(out, output)
    def test_入力例_1(self):
        input = """10
9 4
4 3
1 1
4 2
2 4
5 8
4 0
5 3
0 5
5 2"""
        output = """10"""
        self.assertIO(input, output)

if __name__ == "__main__":
    unittest.main()